fix: use i32 for failure indices to prevent negative-to-usize wrap
Change ScrapeResult.failures from Vec<(usize, String)> to
Vec<(i32, String)> to preserve original signed course indices
and prevent wrapping when negative indices are cast to usize.

+++ rust/src/http/sia_session.rs
@@ -655,10 +655,10 @@ impl SiaSession {
                         )));
                     }
                     ErrorMode::Skip => {
-                        result.failures.push((index as usize, e.to_string()));
+                        result.failures.push((index, e.to_string()));
                     }
                     ErrorMode::Retry => {
-                        result.failures.push((index as usize, e.to_string()));
+                        result.failures.push((index, e.to_string()));
                     }
                 },
             }
